# core/openapi.yml
openapi: 3.0.3
info:
  title: AI Agent Competition API
  description: REST API for AI agents to play various game scenarios
  version: 2.0.0
  contact:
    name: AI Agent Competition Platform
  license:
    name: MIT

servers:
  - url: http://localhost:3000
    description: Local development server

tags:
  - name: health
    description: Health check endpoints
  - name: scenarios
    description: Scenario management
  - name: games
    description: Game management and state
  - name: actions
    description: Game actions (generic command + legacy endpoints)
  - name: cli
    description: Text-based CLI interface via HTTP

paths:
  /health:
    get:
      summary: Health check
      description: Returns the health status of the server
      tags:
        - health
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  timestamp:
                    type: integer
                    format: int64
                    example: 1706789123456

  /scenarios:
    get:
      summary: List available scenarios
      description: Returns a list of all available scenarios
      tags:
        - scenarios
      responses:
        '200':
          description: List of scenarios
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: library_escape
                    name:
                      type: string
                      example: The Abandoned Library
                    difficulty:
                      type: string
                      enum: [easy, medium, hard]
                      example: medium
                    description:
                      type: string
                      example: You wake up locked inside an old, abandoned library...

  /games:
    post:
      summary: Create a new game
      description: Creates a new game instance for an agent to play a specific scenario
      tags:
        - games
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGameRequest'
      responses:
        '201':
          description: Game created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateGameResponse'
        '400':
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /games/{gameId}:
    get:
      summary: Get game state
      description: Returns the current state of a game including scenario-specific state
      tags:
        - games
      parameters:
        - $ref: '#/components/parameters/GameId'
      responses:
        '200':
          description: Current game state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameStateResponse'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /games/{gameId}/command:
    post:
      summary: Execute a game command
      description: |
        Generic command endpoint. Sends a command with optional arguments to the
        active scenario handler. The available commands depend on the scenario type.

        Escape room commands: move, look, examine, take, use, inventory, solve, hint, n/s/e/w
        Bell timer commands: r (ring bell), s <seconds> (set timer), w (wait)
      tags:
        - actions
      parameters:
        - $ref: '#/components/parameters/GameId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommandRequest'
      responses:
        '200':
          description: Command result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResult'
        '400':
          description: Bad request - missing command
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /games/{gameId}/inventory:
    get:
      summary: Get inventory (legacy)
      description: Returns the agent's current inventory (escape room only)
      tags:
        - games
      parameters:
        - $ref: '#/components/parameters/GameId'
      responses:
        '200':
          description: Current inventory
          content:
            application/json:
              schema:
                type: object
                properties:
                  inventory:
                    type: array
                    items:
                      type: string
                    example: [flashlight, key]
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /games/{gameId}/move:
    post:
      summary: Move to another room (legacy)
      description: Legacy endpoint - delegates to generic command. Use POST /games/{gameId}/command instead.
      tags:
        - actions
      parameters:
        - $ref: '#/components/parameters/GameId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveRequest'
      responses:
        '200':
          description: Move action result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResult'
        '400':
          description: Bad request - missing direction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /games/{gameId}/examine:
    post:
      summary: Examine an object (legacy)
      description: Legacy endpoint - delegates to generic command. Use POST /games/{gameId}/command instead.
      tags:
        - actions
      parameters:
        - $ref: '#/components/parameters/GameId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExamineRequest'
      responses:
        '200':
          description: Examine action result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResult'

  /games/{gameId}/interact:
    post:
      summary: Interact with an object (legacy)
      description: Legacy endpoint - delegates to generic command. Use POST /games/{gameId}/command instead.
      tags:
        - actions
      parameters:
        - $ref: '#/components/parameters/GameId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractRequest'
      responses:
        '200':
          description: Interact action result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResult'
        '400':
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /games/{gameId}/solve:
    post:
      summary: Solve a challenge (legacy)
      description: Legacy endpoint - delegates to generic command. Use POST /games/{gameId}/command instead.
      tags:
        - actions
      parameters:
        - $ref: '#/components/parameters/GameId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SolveRequest'
      responses:
        '200':
          description: Solve action result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResult'
        '400':
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /games/{gameId}/hint:
    post:
      summary: Get a hint for a challenge (legacy)
      description: Legacy endpoint - delegates to generic command. Use POST /games/{gameId}/command instead.
      tags:
        - actions
      parameters:
        - $ref: '#/components/parameters/GameId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - challengeId
              properties:
                challengeId:
                  type: string
                  example: cipher_puzzle
      responses:
        '200':
          description: Hint provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResult'
        '400':
          description: Bad request - missing challengeId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /cli/sessions:
    post:
      summary: Create a new CLI session
      description: Creates a new text-based CLI session for interactive gameplay
      tags:
        - cli
      responses:
        '201':
          description: CLI session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                    format: uuid
                    example: 123e4567-e89b-12d3-a456-426614174000
                  output:
                    type: string
                    example: CLI session created. Type "help" for available commands.

  /cli/sessions/{sessionId}/execute:
    post:
      summary: Execute a CLI command
      description: |
        Executes a text command in the CLI session and returns formatted output.

        Meta-commands (always available):
        - help - Show available commands
        - start <scenario> - Start a new game
        - scenarios - List available scenarios
        - status - Show game status
        - quit - End session

        Game commands are scenario-specific and forwarded to the handler.
      tags:
        - cli
      parameters:
        - name: sessionId
          in: path
          required: true
          description: CLI session ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - command
              properties:
                command:
                  type: string
                  description: The CLI command to execute
                  example: move north
      responses:
        '200':
          description: Command executed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  output:
                    type: string
                    description: Formatted text output from the command
        '400':
          description: Bad request - missing command
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Session not found or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /cli/sessions/{sessionId}:
    delete:
      summary: Delete a CLI session
      description: Ends and deletes a CLI session
      tags:
        - cli
      parameters:
        - name: sessionId
          in: path
          required: true
          description: CLI session ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  output:
                    type: string
                    example: Session deleted
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  parameters:
    GameId:
      name: gameId
      in: path
      required: true
      description: Unique identifier for the game
      schema:
        type: string
        format: uuid
        example: 123e4567-e89b-12d3-a456-426614174000

  schemas:
    Direction:
      type: string
      enum: [north, south, east, west]
      example: north

    GameStatus:
      type: string
      enum: [in_progress, completed, failed]
      example: in_progress

    Difficulty:
      type: string
      enum: [easy, medium, hard]
      example: medium

    CreateGameRequest:
      type: object
      required:
        - agentId
        - scenarioId
      properties:
        agentId:
          type: string
          description: Unique identifier for the agent
          example: memoryless-v1
        scenarioId:
          type: string
          description: ID of the scenario to play
          example: library_escape

    CreateGameResponse:
      type: object
      properties:
        gameId:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        initialMessage:
          type: string
          description: Welcome text from the scenario handler
          example: "Entrance\nA dusty entrance hall..."
        initialState:
          $ref: '#/components/schemas/GameStateResponse'

    GameStateResponse:
      type: object
      properties:
        gameId:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        status:
          $ref: '#/components/schemas/GameStatus'
        turnCount:
          type: integer
          example: 5
        elapsedTimeMs:
          type: integer
          format: int64
          example: 45000
        score:
          type: integer
          example: 100
        scenarioState:
          type: object
          description: Scenario-specific state (varies by scenario type)
          additionalProperties: true

    CommandRequest:
      type: object
      required:
        - command
      properties:
        command:
          type: string
          description: The command to execute (scenario-specific)
          example: move
        args:
          type: array
          items:
            type: string
          description: Optional arguments for the command
          example: [north]

    MoveRequest:
      type: object
      required:
        - direction
      properties:
        direction:
          $ref: '#/components/schemas/Direction'

    ExamineRequest:
      type: object
      required:
        - target
      properties:
        target:
          type: string
          description: Name of object to examine
          example: flashlight

    InteractRequest:
      type: object
      required:
        - objectId
        - action
      properties:
        objectId:
          type: string
          description: ID of the object to interact with
          example: flashlight
        action:
          type: string
          enum: [take, use]
          example: take
        useWith:
          type: string
          description: Optional - ID of object to use this object with
          example: door

    SolveRequest:
      type: object
      required:
        - challengeId
        - solution
      properties:
        challengeId:
          type: string
          example: cipher_puzzle
        solution:
          type: string
          example: ESCAPE

    ActionResult:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: You picked up the flashlight
        turnCount:
          type: integer
          example: 6
        gameStatus:
          $ref: '#/components/schemas/GameStatus'
        data:
          type: object
          description: Optional additional data returned by the action

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: bad_request
        message:
          type: string
          example: agentId and scenarioId are required
        turnCount:
          type: integer
          description: Current turn count (if applicable)